Implement hap_persist for esp8266 spiflash
- Rename project to 'hapd'
- Add license file

This project was intended for embedded devices, but it turns out that
the code is consuming too much memory that the esp8266 don't even
have enough heap to connect to Wi-Fi (not to mention pairings). The
next step will be to optimize the code structure: use stacks instead
of heaps to avoid potential memory leaks, remove some async code to
prevent overheads, or start from ground again if there is no other
way.

My original purposes for this project were not only to discover more
about Apple's HomeKit Accessory Protocol, but also to have fun making
things work. And it seems like I have almost archived both goals. So
the next thing is to make some truely working & easy to work with
libraries for other people. Right now the library works just fine on
native platforms (tested on macOS HS and ubuntu 18.04).